requirements.txt
amqp==5.1.1
# aniso8601==7.0.0  # ✅ REMOVED: Not used - was GraphQL dependency
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs
bcrypt==5.0.0
# beautifulsoup4==4.14.2  # ✅ REMOVED: Not used - lxml is sufficient
billiard==4.1.0
bleach
# braintree==4.40.0  # ✅ REMOVED: Payment gateway not used
# Brotli==1.1.0  # ✅ REMOVED: Compression library not used
cachetools==6.0.0
celery==5.3.4
certifi==2024.7.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.3
# confuse==2.0.1  # ✅ REMOVED: Not used in project
# contextlib2==21.6.0  # ✅ REMOVED: Built-in in Python 3.7+
coreapi==2.3.3
coreschema==0.0.4
# crispy-bootstrap4==2024.10  # ✅ REMOVED: Using bootstrap5 only
crispy-bootstrap5==2025.6
# crispy-tailwind==0.5.0  # ✅ REMOVED: Project uses Bootstrap, not Tailwind
# cron_descriptor==1.4.0  # ✅ REMOVED: Not used in codebase
cryptography==46.0.3  # Already compatible with Python 3.13
# cssselect2  # ✅ REMOVED: WeasyPrint dependency - replaced with ReportLab
defusedxml==0.7.1
Deprecated==1.2.18
diff-match-patch==20241021
dill==0.4.0
dj-database-url==2.3.0
Django==4.2.26
# django-allauth==65.7.0  # ✅ REMOVED: Not in INSTALLED_APPS, not used
django-appconf==1.1.0
# django-autoslug==1.9.9  # ✅ REMOVED: Not used in project
django-bleach==3.1.0
django-bootstrap-datepicker-plus==5.0.6
# django-bootstrap4==25.2  # ✅ REMOVED: Using bootstrap5 only
django-bootstrap5==25.2
# django-ckeditor==6.7.3  # ✅ REMOVED: Not in INSTALLED_APPS
# django-ckeditor-5==0.2.18  # ✅ REMOVED: Not in INSTALLED_APPS
django-colorfield==0.14.0
django-compressor==4.5.1
django-cors-headers==4.3.1
django-countries==8.1.1
django-crispy-forms==2.4
django-environ==0.11.2
# django-file-form==3.5.2  # ✅ REMOVED: Not used in project
django-filter==25.1
django-import-export==4.3.13
# django-jet-reboot==1.3.5  # ✅ REMOVED: Not in INSTALLED_APPS
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-model-utils==4.3.1
# django-modeltranslation==0.18.11  # ✅ REMOVED: Not used in project
django-mptt==0.18.0
django-ratelimit==4.1.0
django-redis==5.4.0
# django-redis-cache==3.0.1  # ✅ REMOVED: Using django-redis instead
# django-rest-framework==0.1.0  # ✅ REMOVED: Duplicate - use djangorestframework instead
django-role-permissions==3.2.0
django-select2==8.2.3
django-simple-history==3.10.1
django-smart-selects==1.7.2
# django-storages==1.14.2  # ✅ REMOVED: Not used - using whitenoise only
# django-summernote==0.8.20.0  # ✅ REMOVED: Not in INSTALLED_APPS
django-tables2==2.7.5
# django-taggit==6.1.0  # ✅ REMOVED: Not used in project
# django-tailwind==3.6.0  # ✅ REMOVED: Project uses Bootstrap, not Tailwind
# django-tinymce==5.0.0  # ✅ REMOVED: Not in INSTALLED_APPS
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
et-xmlfile==1.1.0
exceptiongroup
filelock==3.17.0
# git-filter-repo==2.47.0  # ✅ REMOVED: CLI tool, not a Python library
# graphql-core==2.3.2  # ✅ REMOVED: Not used - no GraphQL API in project
greenlet==2.0.2
gunicorn==21.2.0
hiredis==2.3.2
# html5lib==1.1  # ✅ REMOVED: Dependency of beautifulsoup4 (removed)
idna==3.7
importlib_metadata
inflection
iniconfig
# ipaddr==2.2.0  # ✅ REMOVED: Built-in as ipaddress in Python 3.3+
# isort==5.13.2  # ✅ REMOVED: Code formatting tool not used in runtime
itypes==1.2.0
jdcal==1.4.1
# Jinja2==3.1.6  # ✅ REMOVED: Only used in scripts/generate_test_report.py - can be replaced with string formatting
jsonschema
jsonschema-specifications
# lockfile==0.12.2  # ✅ REMOVED: Not used in project
lxml==4.9.3
Markdown==3.4.3
# MarkupPy==1.14  # ✅ REMOVED: Not used in project
MarkupSafe==3.0.2
# matplotlib==3.8.2  # ✅ REMOVED: Optional dependency for stress testing only
# mccabe==0.7.0  # ✅ REMOVED: Linting tool not used in runtime
msgpack==1.0.5
numpy==2.0.2  # ✅ UPDATED: From 1.26.2 to 2.0.2 for Python 3.13 compatibility
# oauthlib==3.2.2  # ✅ REMOVED: django-allauth dependency (allauth not used)
openpyxl==3.1.2
pandas==2.2.3  # ✅ UPDATED: From 2.1.4 to 2.2.3 for Python 3.13 compatibility
# paramiko==4.0.0  # ✅ REMOVED: Only used in deploy.py - can use subprocess/ssh instead
pathspec==0.12.1
# paypalrestsdk==1.13.3  # ✅ REMOVED: PayPal not used
# pep517==0.13.0  # ✅ REMOVED: Build tool not needed in runtime
pillow==11.0.0  # ✅ UPDATED: From 10.3.0 to 11.0.0 for Python 3.13 compatibility
platformdirs
pluggy
# progress==1.6  # ✅ REMOVED: Not used in codebase
psutil==7.1.3  # ⚠️ KEEP: Used in tests/performance, core/views, client/cache_configuration
pytest==7.4.4
pytest-django==4.8.0
pytest-cov==4.0.0
# promise==2.3  # ✅ REMOVED: Dependency of graphql-core (removed)
# pycodestyle==2.10.0  # ✅ REMOVED: Linting tool not used in runtime
pycparser==2.21
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.7.0
PyMySQL==1.1.1
PyNaCl==1.6.1
pyOpenSSL==25.3.0
pyparsing==3.1.0
pypng==0.20220715.0
python-barcode==0.15.1
python-bidi==0.4.2
python-crontab==2.7.1
python-dateutil==2.8.2
# python-decouple==3.8  # ✅ REMOVED: Duplicate - using django-environ instead
python-dotenv==1.0.0  # ⚠️ KEEP: Used in setup_development.py
python-json-logger==2.0.7
# python-magic==0.4.27  # ✅ REMOVED: Not used in codebase
# python-crontab==2.7.1  # ✅ REMOVED: Not used - using celery.schedules.crontab instead
python-slugify==8.0.1
# python3-openid==3.2.0  # ✅ REMOVED: django-allauth dependency (allauth not used)
# pytoml==0.1.21  # ✅ REMOVED: Use toml instead (already in requirements)
pytz==2023.3
PyYAML==6.0.3
pyzk==0.9
qrcode==7.4.2
RapidFuzz==3.13.0
rcssmin==1.1.2
redis==5.2.1  # ✅ UPDATED: From 3.5.3 to latest stable
referencing
regex==2023.6.3
requests
# requests-oauthlib==1.3.1  # ✅ REMOVED: django-allauth dependency (allauth not used)
# retrying==1.3.4  # ✅ REMOVED: Not used - can implement simple retry if needed
rjsmin==1.2.2
rpds-py
# Rx==1.6.3  # ✅ REMOVED: Dependency of graphql-core (removed)
# schedule==1.2.0  # ✅ REMOVED: Only used in bridge_agent - use celery beat instead
sentry-sdk==2.8.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.5.0
# sslcommerz-lib==1.0  # ✅ REMOVED: SSLCommerz payment gateway not used
# stripe==5.5.0  # ✅ REMOVED: Stripe payment gateway not used
text-unidecode==1.3
# tinycss2  # ✅ REMOVED: WeasyPrint dependency - replaced with ReportLab
# tinyhtml5==2.0.0  # ✅ REMOVED: Not used in project
toml==0.10.2
tomli
tomlkit
typing-inspection==0.4.2
typing_extensions
tzdata==2025.2
tzlocal==5.0.1
uritemplate==4.2.0
uritools==4.0.1
urllib3==2.2.2
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.5.0
wrapt==1.17.2
xlsxwriter==3.1.9